+++ samples/wifi_ble_coex_demo/src/main.c
@@ -426,7 +426,7 @@ int bt_throughput_test_init(void)
 	return 0;
 }
 
-static void dhcp_callback_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void dhcp_callback_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 				  struct net_if *iface)
 {
 	int i = 0;
@@ -730,7 +730,7 @@ static void handle_wifi_connect_result(struct net_mgmt_event_callback *cb)
 	k_sem_give(&wlan_sem);
 }
 
-static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 				    struct net_if *iface)
 {
 	switch (mgmt_event) {
